evaluation/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
psutil>=5.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
